+++ ModelsStore/DTO/TABLES/NOTAS_CREDITO.cs
@@ -9,7 +9,7 @@ namespace ModelsStore.DTO.TABLES
     public class NOTAS_CREDITO
     {
         public string CODIGO_NOTA { get; set; }
-        public float VALOR_NOTA { get; set; }
+        public Int64 VALOR_NOTA { get; set; }
         public string DESCRIPCION_NOTA { get; set; }
